+++ src/main/kotlin/omock/model/stonestate/placed/Black.kt
@@ -1,4 +1,4 @@
-package omock.model.stonestate
+package omock.model.stonestate.placed
 
 import omock.model.stone.Stone
