+++ src/main/kotlin/omock/model/stonestate/placed/White.kt
@@ -1,5 +1,6 @@
-package omock.model.stonestate
+package omock.model.stonestate.placed
 
 import omock.model.stone.Stone
+import omock.model.stonestate.placed.Placed
 
 class White(private val stone: Stone) : Placed(stone)
