+++ src/main/kotlin/omock/model/stonestate/unplaced/Clear.kt
@@ -1,6 +1,6 @@
-package omock.model.stonestate
+package omock.model.stonestate.unplaced
 
 import omock.model.stone.Stone
-import omock.model.stonestate.UnPlaced
+import omock.model.stonestate.unplaced.UnPlaced
 
 class Clear(private val stone: Stone) : UnPlaced(stone)
