+++ src/test/kotlin/omock/StoneStateTest.kt
@@ -1,12 +1,12 @@
 package omock
 
-import omock.model.stonestate.Black
+import omock.model.stonestate.placed.Black
 import omock.model.player.BlackPlayer
-import omock.model.stonestate.Clear
+import omock.model.stonestate.unplaced.Clear
 import omock.model.position.Column
 import omock.model.position.Row
 import omock.model.stone.Stone
-import omock.model.stonestate.White
+import omock.model.stonestate.placed.White
 import omock.model.player.WhitePlayer
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
